requirements.txt
# requirements.txt
"""
streamlit>=1.28.0
streamlit-option-menu>=0.3.6
beautifulsoup4>=4.12.0
requests>=2.31.0
aiohttp>=3.9.0
asyncio>=3.4.3
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # or sqlite3 (built-in)
lxml>=4.9.0
networkx>=3.0
plotly>=5.17.0
joblib>=1.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
redis>=5.0.0  # for queue management
celery>=5.3.0  # for distributed tasks
uvicorn>=0.24.0
fastapi>=0.104.0  # for API endpoints
streamlit-autorefresh>=0.1.6
streamlit-extras>=0.4.0
scipy>=1.11.0
tqdm>=4.65.0
rich>=13.0.0
python-magic>=0.4.0
dateparser>=1.1.0
wordcloud>=1.9.0
streamlit-ace>=0.1.1
"""